% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time.R
\name{convert_time}
\alias{convert_time}
\title{Convert Time Units}
\usage{
convert_time(x, from = NULL, to = NULL, warn_residual = TRUE)
}
\arguments{
\item{x}{Numeric value representing the time duration or interval.}

\item{from}{Character string specifying the time unit of the input \code{x}.
Allowed values: "years", "months", "days", "hours", "minutes", "seconds".}

\item{to}{Character vector specifying the desired time units for the output.
Allowed values: "years", "months", "days", "hours", "minutes", "seconds".}

\item{warn_residual}{logical. If TRUE, displays a warning message when there's a residual time not displayed in the output units. Default is TRUE.}
}
\value{
Character string representing the time value in the specified time units.
}
\description{
Convert a given time value from one time unit to another.
}
\details{
This function converts a given time value \code{x} from one time unit to another,
         as specified by the \code{from} and \code{to} parameters. It handles conversions
         between various time units, such as years, months, days, hours, minutes, and seconds.
         The output string includes the numeric amount and the first letter of the time unit.
}
\examples{
convert_time(9000, "seconds", c("days", "hours"))  # Output: "2d 30h"
convert_time(1.22, "years", c("days", "hours", "minutes"))  # Output: "445d 14h 48m"
convert_time(1.22, "years", "months")  # Output: "14m"
}
\seealso{
\code{\link{lubridate::pretty}} for an alternative method to display time durations.
}
