% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time_units.R
\name{convert_time_units}
\alias{convert_time_units}
\title{Convert Time Units}
\usage{
convert_time_units(time_data, from_unit, to_unit, output_format = "character")
}
\arguments{
\item{time_data}{Numeric vector containing the time values to be converted.}

\item{from_unit}{Character string indicating the original time unit of the \code{time_data}.}

\item{to_unit}{Character string indicating the target time unit to which the \code{time_data} should be converted.}

\item{output_format}{Character string specifying the desired output format. Possible values are "character" (default), "period", or "duration".}
}
\value{
Depending on the \code{output_format}, the function returns either a character vector with formatted time values, a 'lubridate' period object, or a 'lubridate' duration object.
}
\description{
Convert time values between different units such as years, months, weeks, days, hours, minutes, and seconds.
}
\details{
This function allows for the conversion of time data between different units, such as years, months, weeks, days, hours, minutes, and seconds. It utilizes the 'lubridate' package to facilitate the conversions. If the package is not installed, it will be automatically installed and loaded upon function call. If the 'lubridate' package is already installed but not loaded, the function will load it.

The function takes a numeric vector of time values (\code{time_data}), the original time unit of the data (\code{from_unit}), and the target time unit for the conversion (\code{to_unit}). The result can be returned as a character string with a formatted time representation ("character"), as a 'lubridate' period object ("period"), or as a 'lubridate' duration object ("duration").

Note: For precise time calculations, this function uses approximate values for the number of seconds in a month and a year.
}
\examples{
convert_time_units(18, "months", "years", output_format = "character")
convert_time_units(181, "months", "years", output_format = "period")
convert_time_units(181, "months", "years", output_format = "duration")

}
\seealso{
\code{\link[lubridate]{period}}, \code{\link[lubridate]{duration}}
}
\author{
Your Name
}
\keyword{conversion}
\keyword{lubridate}
\keyword{time}
