% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_previous_weekday.R
\name{difftime_previous_weekday}
\alias{difftime_previous_weekday}
\title{Calculate the Difference in Days to the Previous Specified Weekday}
\usage{
difftime_previous_weekday(target_date, weekday_name)
}
\arguments{
\item{target_date}{A date object or a character string representing the starting date.}

\item{weekday_name}{A character string representing the name of the desired previous weekday (e.g., "samedi" for Saturday in French).}
}
\value{
An integer representing the number of days difference between the target_date and the nearest previous occurrence of the specified weekday.
}
\description{
This function calculates the number of days between a given date and the nearest previous occurrence of a specified weekday.
}
\examples{
difftime_previous_weekday("2019-07-03", "samedi")
difftime_previous_weekday(as.Date("2021-12-15"), "mardi")

}
