% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_package_functions.R
\name{export_package_functions}
\alias{export_package_functions}
\title{Alias for \code{dump_package_functions()}}
\usage{
export_package_functions(package_name, path, only_exported = TRUE)
}
\description{
Alias for \code{dump_package_functions()}
}
\seealso{
\code{\link[=dump_package_functions]{dump_package_functions()}}
}
