% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffreq.R
\name{ffreq}
\alias{ffreq}
\title{ffreq: Function to compute frequency counts and percentages}
\usage{
ffreq(data, vars, group_vars = NULL, na.rm = F)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{vars}{A character vector of variable names for which to compute frequency counts and percentages.}

\item{group_vars}{A character vector of variable names to group by (optional). Default is NULL.}

\item{na.rm}{A logical value indicating whether to remove NA values in the grouping variable(s). Default is TRUE.}
}
\value{
A data frame with frequency counts and percentages. If `group_vars` is provided, the output is grouped by these variables. The resulting data frame has columns for each variable, their counts, frequencies by group, and absolute frequencies.
}
\description{
This function calculates the frequency counts and percentages for specified columns in a data frame, optionally grouped by one or more grouping variables.
}
\examples{
# Sample data
data <- tibble(
  PATSEX = rep(c("F", "M"), each = 10),
  migration_lithiasique = sample(c("Yes", "No"), 20, replace = TRUE),
  anticoagulants = sample(c("Yes", "No"), 20, replace = TRUE)
)

# vars of interest
vars <- c("migration_lithiasique", "anticoagulants")

# With grouping variable
result_with_group <- ffreq(data, vars, group_vars = "PATSEX")
print(result_with_group)

# Without grouping variable
result_without_group <- ffreq(data, vars)
print(result_without_group)

}
