% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Find the mode(s) of a vector}
\usage{
find_mode(x)
}
\arguments{
\item{x}{A numeric or character vector.}
}
\value{
A vector containing the mode(s) of the input vector. If the input vector has multiple modes, all will be returned.
}
\description{
This function calculates the mode (most frequently occurring value) in a numeric or character vector.
In case of multiple modes, all modes are returned.
}
\examples{
find_mode(c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4))
find_mode(c("a", "b", "b", "c", "c", "c", "d", "d", "d", "d"))

}
