% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_colnames.R
\name{fix_colnames}
\alias{fix_colnames}
\title{Fix column names of a dataframe}
\usage{
fix_colnames(df)
}
\arguments{
\item{df}{A dataframe whose column names need to be fixed.}
}
\value{
A dataframe with fixed column names.
}
\description{
This function trims whitespace, replaces dots with underscores,
transforms to lowercase, and removes diacritic characters from
the column names of a given dataframe.
}
\examples{
df <- data.frame("A .column " = 1:5, "b.Column" = 6:10, check.names = FALSE)
fix_colnames(df)

}
