% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_metrics.R
\name{functions_in_file}
\alias{functions_in_file}
\title{List top-level functions defined in a single R file (static; no sourcing)}
\usage{
functions_in_file(path, as_table = FALSE)
}
\arguments{
\item{path}{Path to a \code{.R} file.}

\item{as_table}{If \code{TRUE}, return a \code{file}/\code{function_name} data frame instead
of a character vector.}
}
\value{
A character vector of function names found.
}
\description{
Parses a file and returns the names of functions defined at the top level,
i.e. expressions of the form \code{name <- function(...) {}} or
\code{assign("name", function(...) {})}. No code is executed.
}
\examples{
\dontrun{
functions_in_file("R/gpt.R")
}
}
