% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_fill_na.R
\name{fuzzy_fill_na}
\alias{fuzzy_fill_na}
\title{Perform a fuzzy join between two dataframes and fill NA values}
\usage{
fuzzy_fill_na(
  df1,
  df2,
  match_cols,
  match_fun_list,
  fill_cols,
  replace_cols = NULL,
  keep_cols = NULL
)
}
\arguments{
\item{df1}{A dataframe. The dataframe to fill NA values in.}

\item{df2}{A dataframe. The dataframe to fill NA values from.}

\item{match_cols}{A character vector. The names of the columns to match on.}

\item{match_fun_list}{A list of functions. The matching functions to use for each column specified in match_cols.
Each function should return a logical vector indicating the matches.}

\item{fill_cols}{A character vector. The names of the columns in df1 to fill NA values in.}

\item{replace_cols}{A character vector, optional. The names of the columns in df1 to replace values in. Default is NULL.}

\item{keep_cols}{A character vector, optional. The names of the columns in df2 to keep in the result. Default is NULL.}
}
\value{
A dataframe. The first dataframe with NA values filled.
}
\description{
This function performs a fuzzy join between two dataframes based on specified matching columns and functions.
It then fills NA values in the specified columns of the first dataframe with the corresponding values from the second dataframe.
}
\details{
This function uses the \code{\link{within_days}} and \code{\link{match_stringdist}} functions for matching.
It has dependencies on the \code{fuzzyjoin}, \code{dplyr}, \code{stringdist}, and \code{lubridate} packages.
}
\examples{
\dontrun{
# Create some synthetic data
df1 <- data.frame(
  name = c("John", "Alice", "Bob", "Charlie"),
  birthdate = as.Date(c("2000-01-01", "2000-02-02", "2000-03-03", "2000-04-04")),
  city = c("New York", "Los Angeles", "Chicago", "San Francisco"),
  id = c(NA, "123", "456", NA),
  stringsAsFactors = FALSE
)
df2 <- data.frame(
  name = c("John", "Alice", "Bob", "Charlie", "Charlie"),
  birthdate = as.Date(c("2000-01-05", "2000-02-02", "2000-03-03", "2000-04-01", "2000-04-07")),
  city = c("New York", "Los Angeles", "Chicago", "San Francisco", "San Francisco"),
  id = c("789", "123", "456", "789", "012"),
  stringsAsFactors = FALSE
)
# Define the matching columns and functions
match_cols <- c("name", "birthdate", "city")
match_fun_list <- list(`stringdist::stringdist`, `\%within\%`, `stringdist::stringdist`)
# Define the columns to fill
fill_cols <- c("id")
# Perform the fuzzy join and fill NA values
df_filled <- fuzzy_fill_na(df1, df2, match_cols, match_fun_list, fill_cols, stringdist_threshold = 2, stringdist_method = "jw")
}
}
