% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeason.R
\name{getSeason}
\alias{getSeason}
\title{Get Season from Date}
\usage{
getSeason(input.date)
}
\arguments{
\item{input.date}{A Date object or a character string representing a date.}
}
\value{
A character string representing the season of the input date: "Winter", "Spring", "Summer", or "Fall".
}
\description{
This function determines the meteorological season (Winter, Spring, Summer, or Fall)
from a given date.
}
\note{
The function uses the following date ranges to determine seasons:
\itemize{
\item Winter: Dec 21 - Mar 19
\item Spring: Mar 20 - Jun 20
\item Summer: Jun 21 - Sep 21
\item Fall: Sep 22 - Dec 20
Dates are treated as being in the northern hemisphere.
}
}
\examples{
getSeason(as.Date("2023-06-15"))
getSeason(as.Date("2023-12-15"))

}
