% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_only_cols.R
\name{na_only_cols}
\alias{na_only_cols}
\title{Get Names of Columns with Only NA Values}
\usage{
na_only_cols(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
A character vector containing the names of the columns where all values are \code{NA}.
}
\description{
This function returns the names of the columns in a data frame or tibble that contain only \code{NA} values.
}
\examples{
df <- data.frame(a = c(1, 2, 3),
                 b = c(NA, NA, NA),
                 c = c(4, NA, 6))
na_only_cols(df)
# Returns: "b"

}
