% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_pyramid.R
\name{plot_age_pyramid}
\alias{plot_age_pyramid}
\title{Plot Age Pyramid}
\usage{
plot_age_pyramid(
  data,
  age_var,
  sex_var,
  breaks = NULL,
  ...,
  install_ggplot2 = TRUE
)
}
\arguments{
\item{data}{A data frame containing the age and sex variables.}

\item{age_var}{The name of the column in the \code{data} data frame representing age.}

\item{sex_var}{The name of the column in the \code{data} data frame representing sex.}

\item{breaks}{The breaks for the \code{cut()} function used to categorize ages into groups.}

\item{...}{Additional arguments passed to ggplot for customization of the plot.
For example, you can use \code{labs()}, \code{theme()}, etc.}

\item{install_ggplot2}{Logical value indicating whether to install and load ggplot2
if it is not already available (default is \code{TRUE}).}
}
\value{
A ggplot2 plot representing the age pyramid.
}
\description{
This function creates an age pyramid plot using ggplot2 for a given data frame
containing age and sex variables.
}
\examples{
\dontrun{
# Sample data
etude <- data.frame(
  age = c(24, 45, 67, 38, 55, 28, 78, 63, 42, 55),
  sex = c("M", "F", "M", "F", "M", "F", "M", "F", "M", "F")
)

# Plot age pyramid with customized breaks
plot_age_pyramid(etude, age_var = "age", sex_var = "sex", breaks = seq(0, 100, 10),
                 fill = "lightblue", color = "black")
}

}
\keyword{age}
\keyword{ggplot2}
\keyword{plot}
\keyword{pyramid}
\keyword{visualization}
