% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_main.R
\name{process_main}
\alias{process_main}
\title{Process PMSI Main Data}
\usage{
process_main(data)
}
\arguments{
\item{data}{A dataframe containing PMSI data.}
}
\value{
A transformed dataframe containing main PMSI data fields.
}
\description{
This function processes the main PMSI data, including admission and discharge dates, patient age and sex, and stay duration.
}
\examples{
data <- data.frame() # Example data
main_data <- process_main(data)
}
