% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_accent.R
\name{rm_accent}
\alias{rm_accent}
\title{Remove Accents from Characters}
\usage{
rm_accent(str, pattern = "all")
}
\arguments{
\item{str}{The character string from which accents are to be removed.}

\item{pattern}{A character vector specifying the types of accents to be removed. This can include
"´" for acute, "`" for grave, "^" for circumflex, "~" for tilde, "¨" for umlaut, "ç" for cedilla,
and "all" for all types of accents. If not specified, all types of accents are removed.}
}
\value{
A character string with the specified accents removed.
}
\description{
This function removes various types of accents (acute, grave, circumflex, tilde, umlaut, and cedilla)
from a character string. The types of accents to be removed can be specified.
}
\examples{
rm_accent("pâté")
rm_accent("naïve")
rm_accent("papá", pattern = "´")

}
