% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_editable_plot.R
\name{save_editable_plot}
\alias{save_editable_plot}
\title{Save an Editable ggplot2 Plot as a PowerPoint Slide}
\usage{
save_editable_plot(plot, filename, path = "../graphs/", extension = ".pptx")
}
\arguments{
\item{plot}{The ggplot2 plot to be saved.}

\item{filename}{A character string specifying the name of the PowerPoint file to be created (without the .pptx extension).}

\item{path}{A character string specifying the directory where the PowerPoint file should be saved.
Defaults is create a new directory called "graphs" under the current working directory which is in accordance with the ProjectTemplate folders arborescence.}

\item{extension}{Currently the only working extension is .pptx

Add outside of a markdown chunk to allow download:
download \href{graphs/boxplot_alat_score_fibrose.pptx}{📥 Télécharger le graphique (PPTX)}}
}
\value{
A PowerPoint file containing the ggplot2 plot as an editable graphic is saved to the specified directory.
}
\description{
This function takes a ggplot2 plot and saves it as an editable graphic in a new PowerPoint file.
The function checks if the necessary packages ("ggplot2", "officer", "rvg") are installed,
and if not, prompts the user to install them.
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(mpg, hp)) +
  geom_point() +
  theme_minimal()
save_editable_plot(p, filename = "my_plot", path = "your_project_name/graphs")
}
[📥 Télécharger le graphique (PPTX)](graphs/boxplot_alat_score_fibrose.pptx)
}
