% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_duplicated.R
\name{slice_duplicated}
\alias{slice_duplicated}
\title{Slice Duplicated Rows Based on Specific Columns}
\usage{
slice_duplicated(data, ...)
}
\arguments{
\item{data}{A data frame or tibble that you want to inspect for duplicate rows.}

\item{...}{One or more unquoted column names that you want to check for duplicates. You can pass multiple columns to check for duplicates based on a combination of those columns.}
}
\value{
A data frame or tibble containing only the duplicated rows based on the specified columns.
}
\description{
This function allows you to select and inspect rows in a data frame that have duplicates based on specific columns.
It groups the data by the specified columns, filters out unique rows, and returns only the duplicated rows.
}
\examples{
# Example dataset
data <- tibble::tibble(
  id = c(1, 2, 2, 3, 4, 4, 4),
  value = c("A", "B", "B", "C", "D", "D", "D"),
  date = as.Date(c("2023-01-01", "2023-01-02", "2023-01-02", "2023-01-03", "2023-01-04", "2023-01-04", "2023-01-04"))
)

# Use the function to select duplicated rows based on the 'id' column
slice_duplicated(data, id)

# Use the function to select duplicated rows based on both 'id' and 'date' columns
slice_duplicated(data, id, date)

}
