% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_package.R
\name{source_package}
\alias{source_package}
\title{Helper function to load packages from .R script package exports. Mostly used in the EDSAN jupyter env}
\usage{
source_package(
  path = "/appli/edsan_datalake/edsan_data/utils/",
  package_name = ""
)
}
\arguments{
\item{path}{Default to where i save exported packages files.}

\item{package_name}{Package to load.}
}
\description{
Helper function to load packages from .R script package exports. Mostly used in the EDSAN jupyter env
}
