% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stay_duration.R
\name{stay_duration}
\alias{stay_duration}
\title{Calculate Total Stay Duration}
\usage{
stay_duration(data, id_col, start_col, end_col)
}
\arguments{
\item{data}{A dataframe containing the stay data.}

\item{id_col}{The name of the column in `data` that identifies each stay event or patient.}

\item{start_col}{The name of the column in `data` representing the start date of each stay.}

\item{end_col}{The name of the column in `data` representing the end date of each stay.}
}
\value{
A dataframe with each unique identifier from `id_col` and the corresponding
        total duration of stays.
}
\description{
Computes the total duration of stays in a hospital or similar settings,
accounting for overlapping and non-overlapping intervals.
}
\examples{
# Assuming 'pmsi_general' is your dataframe and it has columns 'evtid' (event ID),
# 'datent' (admission date), and 'datsort' (discharge date):
# total_durations <- stay_duration(pmsi_general, 'evtid', 'datent', 'datsort')

}
