% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_days.R
\name{within_days}
\alias{within_days}
\title{Check if two dates are within a certain number of days of each other}
\usage{
within_days(x, y, days = 2)
}
\arguments{
\item{x}{A Date. The first date to compare.}

\item{y}{A Date. The second date to compare.}

\item{days}{An integer. The maximum number of days apart the dates can be.}
}
\value{
A logical. TRUE if the dates are within the specified number of days of each other, FALSE otherwise.
}
\description{
Check if two dates are within a certain number of days of each other
}
