% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wowhead_get_icon_from_spell_id.R
\name{wowhead_get_icon_from_spell_id}
\alias{wowhead_get_icon_from_spell_id}
\title{Get the Wowhead icon URL for a given spell ID}
\usage{
wowhead_get_icon_from_spell_id(spell_id)
}
\arguments{
\item{spell_id}{A character or numeric spell ID (e.g., \code{"47540"} or \code{47540})}
}
\value{
A character string containing the full URL to the icon image
(e.g., \code{"https://wow.zamimg.com/images/wow/icons/large/spell_holy_penance.jpg"}),
or \code{NA} if the icon could not be found or the request failed.
}
\description{
This function scrapes the spell page on Wowhead and extracts the
true in-game icon using the \code{iconFilename} specified in the
\code{WeakAuraExport.setOptions(...)} JavaScript block.
}
\examples{
wowhead_get_icon_from_spell_id(47540)
#> [1] "https://wow.zamimg.com/images/wow/icons/large/spell_holy_penance.jpg"

}
