% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wowhead_get_unit_location.R
\name{wowhead_get_unit_location}
\alias{wowhead_get_unit_location}
\title{Retrieve Locations of NPCs from Wowhead}
\usage{
wowhead_get_unit_location(npc_ids)
}
\arguments{
\item{npc_ids}{A vector of NPC IDs (numeric or character) to retrieve locations for. Each ID corresponds to an NPC on Wowhead.}
}
\value{
A data frame with two columns:
  \describe{
    \item{npc_id}{The NPC ID provided as input.}
    \item{location}{The location(s) where the NPC can be found. If multiple locations exist, they are concatenated as a single string separated by "; ". If no location is found, this will be `NA`.}
  }
If an NPC ID does not exist or if Wowhead fails to load, `location` will contain `NA`.
}
\description{
This function retrieves the in-game locations of multiple World of Warcraft NPCs (mobs) from Wowhead based on their NPC IDs.
}
\details{
This function uses web scraping to retrieve location data from Wowhead’s NPC pages. If an NPC ID is invalid or Wowhead does not provide location information, `location` will be set to `NA`. Note that this function may be slow for a large number of NPC IDs, as it sends a separate request for each ID.
}
\examples{
\dontrun{
# Retrieve location for a single NPC
wowhead_get_unit_location(135263)

# Retrieve locations for multiple NPCs
npc_ids <- c(214840, 135263, 214842)  # Replace with actual NPC IDs
npc_locations <- wowhead_get_unit_location(npc_ids)
print(npc_locations)
}

}
