% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time.R
\name{convert_time}
\alias{convert_time}
\title{Convert time units}
\usage{
convert_time(x, from = NULL, to = NULL, warn_residual = TRUE)
}
\arguments{
\item{x}{A numeric vector, the time durations to be converted.}

\item{from}{A character string, the unit of the input time durations. It should be one of "seconds", "minutes", "hours", "days", "months", or "years".}

\item{to}{A character vector, the units to which the input time durations should be converted. It should contain one or more of "seconds", "minutes", "hours", "days", "months", or "years".}

\item{warn_residual}{A logical value, if TRUE, a warning is printed when there is a residual part not being displayed due to the chosen 'to' units.}
}
\value{
A character vector of the converted time durations. Each element is a string that concatenates the converted time durations in the units specified by 'to', in the order they are given.
}
\description{
This function converts a time duration from one unit to one or more other units.
The conversion is performed separately for each element in the input vector, and the results are returned as a vector of character strings.
The input can contain NA values, which are returned as NA in the output.
}
\examples{
convert_time(90, from = "seconds", to = c("minutes", "seconds")) # "1m 30s"
convert_time(c(90, 3600), from = "seconds", to = c("hours", "minutes", "seconds")) # "0h 1m 30s" "1h 0m 0s"
}
