% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time.R
\name{convert_time}
\alias{convert_time}
\title{Convert Time Units}
\usage{
convert_time(x, from = NULL, to = NULL)
}
\arguments{
\item{x}{Numeric value representing the time duration.}

\item{from}{Character string indicating the starting time unit.
Possible values: "years", "months", "days", "hours", "minutes", "seconds".}

\item{to}{Character vector indicating the desired time units for conversion.
Possible values: "years", "months", "days", "hours", "minutes", "seconds".}
}
\value{
Character string representing the converted time duration in the specified units.
}
\description{
Converts a given time value from one unit to another.
}
\details{
This function converts a given time value 'x' from one time unit to another.
         The 'from' parameter specifies the starting time unit of 'x', and the 'to'
         parameter specifies the desired time units for the conversion. The function
         returns a character string representing the converted time duration in the
         specified units.
}
\note{
The 'from' and 'to' parameters must be specified to avoid any possible mistakes.
      If any of them are missing, the function will throw an error message.
}
\examples{
convert_time(9000, from = "seconds", to = c("hours", "minutes"))
# Returns: "2h 30m"

convert_time(1.22, from = "years", to = c("days", "hours", "minutes"))
# Returns: "445d 14h 48m"

}
