% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time_units.R
\name{convert_time_units}
\alias{convert_time_units}
\title{Convert Time Units}
\usage{
convert_time_units(time_data, from_unit, to_unit, output_format = "character")
}
\arguments{
\item{time_data}{A numeric vector representing the time values to be converted.}

\item{from_unit}{A character string specifying the unit of the input \code{time_data}. Allowed values: "seconds", "minutes", "hours", "days", "weeks", "months", "years".}

\item{to_unit}{A character string specifying the desired output unit. Allowed values: "seconds", "minutes", "hours", "days", "weeks", "months", "years".}

\item{output_format}{A character string specifying the desired output format. Allowed values: "character" (default), "period", or "duration".}
}
\value{
A character string, lubridate period, or duration representing the converted time units, based on the \code{output_format}.
}
\description{
This function converts time units between different scales such as seconds, minutes, hours, days, weeks, months, and years.
}
\details{
The function provides three different output formats:
\itemize{
  \item "character" (default): Returns the result as a character string in the format "x unit y months", where "x" is the whole number of units and "y" is the fractional part in months.
  \item "period": Returns the result as a lubridate period object representing the converted time units.
  \item "duration": Returns the result as a lubridate duration object representing the converted time units.
}
}
\examples{
# Convert 182 months to years and months in character format
convert_time_units(182, "months", "years")

# Convert 182 months to years and months as a lubridate period object
convert_time_units(182, "months", "years", output_format = "period")

# Convert 182 months to years and months as a lubridate duration object
convert_time_units(182, "months", "years", output_format = "duration")

}
\keyword{conversion}
\keyword{duration}
\keyword{lubridate}
\keyword{period}
\keyword{time}
\keyword{units}
