% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_package_functions.R
\name{export_package_functions}
\alias{export_package_functions}
\title{Export Functions from a Package}
\usage{
export_package_functions(package_name, output_file_path)
}
\arguments{
\item{package_name}{A string specifying the name of the package from which to export functions.}

\item{output_file_path}{A string specifying the path and name of the output file where the functions will be written.}
}
\value{
Invisible. The function writes to a file and does not return a value.
}
\description{
This function exports all the functions from a specified R package into a single script file.
Each function is extracted and written to the file with its name, making it easy to source in another environment.
}
\note{
This function only exports R functions and does not handle dependencies or compiled code.
Ensure that the package's license allows for this kind of use.
}
\examples{
export_package_functions("dplyr", "dplyr_functions.R")

}
