% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_colnames.R
\name{fix_colnames}
\alias{fix_colnames}
\title{Fix column names of a dataframe}
\usage{
fix_colnames(df)
}
\arguments{
\item{df}{A dataframe whose column names need to be "fixed".}
}
\value{
A dataframe with fixed column names.
}
\description{
This function take a \emph{df} as argument and return a \emph{df} with modified \emph{colnames} to make it easier to work with
variable names. Here's what it does: trims whitespace, replaces punctuation with underscores, transforms to lowercase,
and removes diacritic characters from the column names of the given dataframe.
}
\examples{
df <- data.frame("A .column " = 1:5, "b.Column" = 6:10, check.names = FALSE)
fix_colnames(df)

}
