% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pmsi.R
\name{process_pmsi}
\alias{process_pmsi}
\title{Wrapper Function to Process PMSI Data}
\usage{
process_pmsi(data)
}
\arguments{
\item{data}{A dataframe containing PMSI data.}
}
\value{
Three dataframes: main PMSI data, actes data, and diagnosis data. The dfs are named "x", "x_doag" and "x_actes".
}
\description{
This function serves as a wrapper to process PMSI data by calling the \code{process_actes}, \code{process_diag}, and \code{process_main} functions.
}
\examples{
data <- data.frame()  # Example data
process_pmsi(data)
}
